Reimplement relative path computation
Don't use strtok to avoid unnecessary allocations and simplify the
code.

Include a test to cover the case where the package and target
directories don't share any common parent (this was a special case in
the original implementation) and setup the path correctly in the
Makefile so that we test the built version, not the one currently
installed.